python/requirements.txt
# Production Dependencies for ARCHON RELOADED
# Generated from pyproject.toml for compatibility with legacy deployment systems
# Primary dependency management should continue using: uv sync --all-extras

fastapi>=0.116.1
httpx>=0.28.1
loguru>=0.7.3
pydantic-settings>=2.10.1
python-dotenv>=1.1.1
python-multipart>=0.0.9
python-socketio[asgi]>=5.13.0
supabase>=2.18.1
tenacity>=9.1.2
uvicorn>=0.35.0
pyjwt>=2.10.1
PyPDF2>=3.0.1
prometheus-client>=0.21.0
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-instrumentation-httpx>=0.57b0

# Sub-dependencies (automatically resolved but listed for transparency)
annotated-types>=0.7.0
anyio>=4.10.0
bidict>=0.23.1
certifi>=2025.8.3
click>=8.2.1
deprecation>=2.1.0
h11>=0.16.0
h2>=4.2.0
hpack>=4.1.0
httpcore>=1.0.9
hyperframe>=6.1.0
idna>=3.10
packaging>=25.0
postgrest>=1.1.1
pydantic>=2.11.7
pydantic-core>=2.33.2
python-dateutil>=2.9.0.post0
python-engineio>=4.12.2
realtime>=2.7.0
simple-websocket>=1.1.0
six>=1.17.0
sniffio>=1.3.1
starlette>=0.47.2
storage3>=0.12.1
strenum>=0.4.15
supabase-auth>=2.12.3
supabase-functions>=0.10.1
typing-extensions>=4.14.1
typing-inspection>=0.4.1
websockets>=15.0.1
wsproto>=1.2.0
